+++ src/Libraries/RevitNodes/Elements/Views/Sheet.cs
@@ -717,12 +717,13 @@ namespace Revit.Elements.Views
         /// Duplicates A Sheet. 
         /// </summary>
         /// <param name="sheet">The Sheet to be Duplicated.</param>
+        /// <param name="duplicateWithContents">Set to true that Duplicate sheet with contents</param>
         /// <param name="duplicateWithView">Set to true that Duplicate sheet with views.</param>
         /// <param name="viewDuplicateOption">Enter View Duplicate Option: 0 = Duplicate. 1 = AsDependent. 2 = WithDetailing.</param>
         /// <param name="prefix"></param>
-        /// <param name="suffix"></param>
+        /// <param name="suffix">When prefix and suffix are both empty, suffix will set a default value - " - Copy".</param>
         /// <returns></returns>
-        public static Sheet DuplicateSheet(Sheet sheet, bool duplicateWithView = false, int viewDuplicateOption = 0, string prefix = "", string suffix = "")
+        public static Sheet DuplicateSheet(Sheet sheet, bool duplicateWithContents = false, bool duplicateWithView = false, int viewDuplicateOption = 0, string prefix = "", string suffix = "")
         {            
             if (sheet == null)
                 throw new ArgumentNullException(nameof(sheet));
@@ -745,7 +746,7 @@ namespace Revit.Elements.Views
 
             if (String.IsNullOrEmpty(prefix) && String.IsNullOrEmpty(suffix))
             {
-                throw new ArgumentNullException(Properties.Resources.SheetDuplicateNeedFix);
+                suffix = " - Copy";
             }
 
             Sheet newSheet = null;
@@ -774,8 +775,16 @@ namespace Revit.Elements.Views
                                 {
                                     newSheet = oldSheet;
                                     TraceElements.AddRange(oldElements);
-                                }                                
-                            }                                
+                                }
+                                if(newSheet != null)
+                                {
+                                    if (duplicateWithContents)
+                                        DuplicateSheetAnnotations(sheet, newSheet);
+                                    else
+                                        DeleteSheetAnnotations(newSheet);
+                                }
+
+                            }
                         }
                     }
                     if(newSheet == null)
@@ -817,14 +826,18 @@ namespace Revit.Elements.Views
                     TraceElements.Add(newSheet.InternalElement);
 
                     // Copy Annotation Elements from sheet to new sheet by ElementTransformUtils.CopyElements
-                    DuplicateSheetAnnotations(sheet, newSheet);
-
-                    // Copy ScheduleSheetInstance except RevisionSchedule from sheet to new sheet by ElementTransformUtils.CopyElements
-                    DuplicateScheduleSheetInstance(sheet, newSheet);
-
-                    // Duplicate Viewport in sheet and place on new sheet
+                    if(duplicateWithContents)
+                        DuplicateSheetAnnotations(sheet, newSheet);
+                    
                     if (duplicateWithView)
+                    {
+                        // Copy ScheduleSheetInstance except RevisionSchedule from sheet to new sheet by ElementTransformUtils.CopyElements
+                        DuplicateScheduleSheetInstance(sheet, newSheet);
+
+                        // Duplicate Viewport in sheet and place on new sheet
                         TraceElements.AddRange(DuplicateViewport(sheet, newSheet, Option, prefix, suffix));
+                    }
+                        
                 }                
                                 
                 ElementBinder.SetElementsForTrace(TraceElements);
@@ -873,13 +886,43 @@ namespace Revit.Elements.Views
                 BuiltInCategory.OST_TextNotes
             };
             List<ElementId> list = new List<ElementId>();
+            List<ElementId> currentList = new List<ElementId>();
             foreach(var category in Filters)
             {
                 list.AddRange(new FilteredElementCollector(Document, oldSheet.InternalElementId).OfCategory(category).ToElementIds());
+                currentList.AddRange(new FilteredElementCollector(Document, newSheet.InternalElementId).OfCategory(category).ToElementIds());
             }
             if(list.Any<ElementId>())
             {
+                if (currentList.Any<ElementId>() && currentList.Count == list.Count)
+                    return;
+                else if (currentList.Any<ElementId>() && currentList.Count != list.Count)
+                    DeleteSheetAnnotations(newSheet);
                 ElementTransformUtils.CopyElements(oldSheet.InternalViewSheet, list, newSheet.InternalViewSheet, null, null);
+            }
+        }
+
+        private static void DeleteSheetAnnotations(Sheet sheet)
+        {
+            List<BuiltInCategory> Filters = new List<BuiltInCategory>
+            {
+                BuiltInCategory.OST_IOSDetailGroups,
+                BuiltInCategory.OST_Dimensions,
+                BuiltInCategory.OST_GenericAnnotation,
+                BuiltInCategory.OST_Lines,
+                BuiltInCategory.OST_TextNotes
+            };
+            List<ElementId> list = new List<ElementId>();
+            foreach (var category in Filters)
+            {
+                list.AddRange(new FilteredElementCollector(Document, sheet.InternalElementId).OfCategory(category).ToElementIds());
+            }
+            if(list.Any<ElementId>())
+            {
+                foreach(var id in list)
+                {
+                    Document.Delete(id);
+                }
             }
         }